Remove MIDI-specific values from the data model; anchor time on Duration
Three sum-type refactors plus the completed ms→Duration migration, all
aimed at keeping synth/MIDI representation out of the core notation model
and confining it to the codec boundary.

Time: wall-clock ms → musical Duration
- PitchedNote/DrumNote and every side-channel anchor at rational Duration
  positions (fractions of a whole note). TempoConversion removed.
- TimeMapper is the sole Duration↔ms translation, integrating TempoTrack.
- JSON wire format serialises Duration as "num/den" strings.

Dynamics: Loudness sum type
- Loudness = Named(Dynamic) | Raw(double level in [0,1]).
- VelocityChange/VolumeChange carry Loudness; dedup is by record equality
  so Named(F) and Raw(0.76) stay distinct and authored marks survive.
- Dynamic enum gains level(); CRESCENDO/DECRESCENDO removed (not levels).
- MusicXmlParser prefers <dynamics><f/> over <sound dynamics="…">, treats
  <staff> as visual anchor not scope — dynamics fan out part-wide.
- Codec maps level→byte (round(level×127)) only at emit/read.

Hairpins: dedicated sidecar
- HairpinSpan(from, to, Direction) carries no level; endpoints resolve
  from the discrete Loudness timeline. Hairpins keyed by TrackId, added
  to Performance with backward-compat ctors.
- Parser pairs <wedge> start/stop by (partId, number).
- HairpinResolver flattens spans to 1/16-note Raw samples as a prelude
  to MidiCodec.toMidiSplit; discrete marks win where they coexist.

Pitch: Pitch sum type
- Pitch = Spelled(Step, alter, octave) | RawMidi(int); Step enum in core.
- PitchedNote canonical takes Pitch; int ctors and midi() kept as
  boundary conveniences so existing call sites compile unchanged.
- Parser emits Spelled verbatim for main and grace notes; RawMidi only
  for transposing parts (no canonical re-spelling without key context).

Also carried on this branch: Lyrics side-channel + notation-lyrics editor,
TimeSignatureTrack/KeySignatureTrack for mid-piece changes, Monophony.

Known gaps: tests in notation-performance/-mxl/-experiments and
RecorderApp still on the old long-ms shape (build with
-Dmaven.test.skip=true); JSON read-back of Loudness/Pitch deduction
untested; no MusicXML write-back yet.

+++ notation-expressivity/src/main/java/music/notation/expressivity/Volume.java
@@ -31,7 +31,7 @@ public record Volume(Map<TrackId, VolumeControl> byTrack) {
 
     public static Volume empty() { return new Volume(Map.of()); }
 
-    public static Volume single(TrackId track, int level) {
+    public static Volume single(TrackId track, double level) {
         return new Volume(Map.of(track, VolumeControl.constant(level)));
     }
 }
